# frinq-mobile/docs/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_upload_voice_api_v1_voice_post": {
        "properties": {
          "audio": {
            "format": "binary",
            "title": "Audio",
            "type": "string"
          },
          "duration_sec": {
            "default": 0,
            "title": "Duration Sec",
            "type": "integer"
          },
          "question_key": {
            "title": "Question Key",
            "type": "string"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "submission_id",
          "question_key",
          "audio"
        ],
        "title": "Body_upload_voice_api_v1_voice_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Frinq Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/voice/{submission_id}": {
      "get": {
        "description": "List voice clips uploaded for a submission. Returns metadata + a\nstreaming URL the admin UI can use as the <audio> src.\n\nReturns empty `clips` + `table_missing: true` if the voice_clips\ntable doesn't exist (instead of HTTP 500), so the admin UI doesn't\nbreak the whole user detail panel when the migration hasn't run yet.",
        "operationId": "list_voice_clips_api_v1_admin_voice__submission_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Voice Clips Api V1 Admin Voice  Submission Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Voice Clips",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/voice": {
      "post": {
        "description": "Accept an audio blob from the voice recorder, upsert into voice_clips.\n\nRe-recording the same question for the same submission overwrites the\nprevious take (ON CONFLICT DO UPDATE). Returns size + ok so the client\ncan confirm the upload succeeded.",
        "operationId": "upload_voice_api_v1_voice_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_voice_api_v1_voice_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Upload Voice Api V1 Voice Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Voice",
        "tags": [
          "voice"
        ]
      }
    },
...
  }
}